Stop printing after calling `stop`
To consistently stop printing after a call to `stop` and it's variants,
we must join the spawned thread. To support the implementation of the
`Drop` Trait, we store the `JoinHandle` as an Option in the `Spinner`
struct. This allows us to explicitly join the thread when the Spinner is
dropped if it hasn't been explicitly stopped, otherwise it does nothing.